+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 ## Unreleased
 
 
+### Added
+
+-   MINOR: new `createMapInternal()` function
+
+-   MINOR: new `createInternal()` export that is same as default export
+
+
 ## 1.0.1 - 2016-12-22